+++ server/src/Domain/Features/FeaturePipelineParameterOrder.cs
@@ -6,7 +6,7 @@ public readonly record struct FeaturePipelineParameterOrder
     public required object Value { get; init; }
 
     public bool? BoolValue => Value as bool?;
-    public int? IntValue => Value as int?;
-    public float? FloatValue => Value as float?;
+    public long? LongValue => Value as long?;
+    public double? DoubleValue => Value as double?;
     public string? StringValue => Value as string;
 }
